Endpoints supabase

+++ Justock-Spring/justock-api/src/main/java/com/justeam/justock_api/repository/OrderRepository.java
@@ -1,9 +1,9 @@
 package com.justeam.justock_api.repository;
 
+import com.justeam.justock_api.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.justeam.justock_api.model.Order;
-
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Long> {}
+public interface OrderRepository extends JpaRepository<Order, Integer> {
+}
